requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coverage==7.0.3
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
python-slugify==7.0.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.13
